+++ manual-tests/test-heatmap.ts
@@ -1,5 +1,5 @@
-import { analyzePGN, printHeatmap } from 'chessalyzer';
-import { TileTracker, type TileTrackerState } from 'chessalyzer/trackers';
+import { analyzePGN, getTrackerState, printHeatmap } from 'chessalyzer';
+import { TileTracker } from 'chessalyzer/trackers';
 
 const tileTracker = new TileTracker();
 const tileTracker2 = new TileTracker();
@@ -21,9 +21,18 @@ const data = await analyzePGN('./pgn/asorted-games.pgn', {
 });
 console.log(data);
 
-const state1 = data.runs[0]?.trackers[0]?.state as TileTrackerState;
-const state2 = data.runs[1]?.trackers[0]?.state as TileTrackerState;
+const state1 = getTrackerState(data, tileTracker);
+const state2 = getTrackerState(data, tileTracker2);
 
-printHeatmap(tileTracker.generateHeatmap(state1, 'PIECE_MOVED_TO_TILE', 'd1'));
-printHeatmap(tileTracker2.generateHeatmap(state2, 'PIECE_MOVED_TO_TILE', 'd1'));
-printHeatmap(tileTracker.generateComparisonHeatmap(state1, state2, 'PIECE_MOVED_TO_TILE', 'd1'));
+printHeatmap(
+    tileTracker.generateHeatmap(state1, { analysis: 'PIECE_MOVED_TO_TILE', square: 'd1' }),
+);
+printHeatmap(
+    tileTracker2.generateHeatmap(state2, { analysis: 'PIECE_MOVED_TO_TILE', square: 'd1' }),
+);
+printHeatmap(
+    tileTracker.generateComparisonHeatmap(state1, state2, {
+        analysis: 'PIECE_MOVED_TO_TILE',
+        square: 'd1',
+    }),
+);