+++ src/index.ts
@@ -1,16 +1,20 @@
 // Runtime
 export { analyzePGN, printHeatmap } from '#core/analyze';
+export { getTrackerState } from '#core/get-tracker-state';
 export { getAnalyzeError, isReplayError } from '#core/analyze-errors';
 
 // Types
 export type {
     AnalyzeMultiRunOptions,
+    AnalyzeMultiRunResult,
     AnalyzeOptions,
     AnalyzeResult,
+    AnalyzeResultBase,
     AnalyzeRun,
     AnalyzeRunResult,
     AnalyzeSharedOptions,
     AnalyzeSingleRunOptions,
+    AnalyzeSingleRunResult,
     GameFilter,
     WorkerOptions,
 } from '#types/analysis';